Remove the PatternStore concept, it's not applicable for our purpose.
