+++ public_html/wp-content/plugins/pattern-translations/pattern-translator.php
@@ -9,7 +9,7 @@ namespace WordPressdotorg\Pattern_Translations;
 
 const GLOTPRESS_PROJECT = 'disabled/patterns';
 
-require __DIR__ . '/pattern-stores.php';
+require __DIR__ . '/pattern.php';
 require __DIR__ . '/parser.php';
 require __DIR__ . '/i18n.php';
 require __DIR__ . '/makepot.php';
